requirements.txt
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
jmespath==1.0.1
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
